feat: one registry for distribution names, and type-check four modules
Closes #383 and #384, which turned out to be the same piece of work.

Names resolved in at least four ways: a table in cli.py, getattr on the
package in performance.py and roadmap.py, and if/elif chains in
validation.py and utilities.py -- five copies of the chain in
validation.py alone. They disagreed about what was acceptable. The CLI
took "student-t" and "gpd"; the chains took "studentt" and had no entry
for the generalized Pareto at all; getattr took the class name and
nothing else. Adding a family meant finding every one of them.

heavytails/registry.py holds it once, with every name any caller
accepted before as an alias, so nothing that worked stops working. The
CLI's table is a view of the registry now rather than a second list, and
a family added to the registry appears in the CLI without a second edit.

That is most of #384 as a side effect. Of the 47 errors the
ignore_errors suppressions were hiding, the majority were the chains
rebinding one variable to five distribution classes; the registry
removed them. The rest were missing annotations, bare dict and tuple
generics, and returns of values a lookup had made opaque.

One was a real inconsistency rather than a missing annotation:
benchmark_sampling declared dict[str, Any] and returns one entry per
sample size, keyed by that size. The annotation was wrong and now says
what the method does.

The registry's error message keeps the wording the lookups it replaces
used -- it opens with "Unknown distribution" because callers match on
that, and three tests in test_performance.py do.

A test compares each family's recorded parameter names against the
constructor signature, and caught GeneralizedPareto listed as
(xi, mu, sigma) where it takes (xi, sigma, mu).

+++ heavytails/extensions.py
@@ -161,7 +161,7 @@ class StudentTCopula(HeavyTailCopula):
         if product_univariate == 0:
             return 0.0
 
-        return multivariate_t_density / product_univariate
+        return float(multivariate_t_density / product_univariate)
 
     def cdf(self, u: list[float]) -> float:
         """
@@ -296,7 +296,7 @@ class ExtremeValueCopula(HeavyTailCopula):
             term = u_val ** (-self.theta) + v_val ** (-self.theta) - 1
             if term <= 0:
                 return 0.0
-            return term ** (-1.0 / self.theta)
+            return float(term ** (-1.0 / self.theta))
 
         elif self.copula_type == "frank":
             # C(u,v) = -1/θ * ln(1 + (e^(-θu)-1)(e^(-θv)-1)/(e^(-θ)-1))
@@ -348,7 +348,7 @@ class ExtremeValueCopula(HeavyTailCopula):
             term3 = A**theta + theta - 1
 
             pdf_val = C * term1 * term2 * term3 / (u_val * v_val)
-            return pdf_val
+            return float(pdf_val)
 
         elif self.copula_type == "clayton":
             # PDF for Clayton copula (derived from CDF)
@@ -364,7 +364,7 @@ class ExtremeValueCopula(HeavyTailCopula):
                 * (u_val * v_val) ** (-1 - theta)
                 * sum_term ** (-1.0 / theta - 2)
             )
-            return pdf_val
+            return float(pdf_val)
 
         elif self.copula_type == "frank":
             # PDF for Frank copula (derived from CDF)
@@ -389,7 +389,7 @@ class ExtremeValueCopula(HeavyTailCopula):
                 return 0.0
 
             pdf_val = numerator / denominator
-            return pdf_val
+            return float(pdf_val)
 
         else:
             raise ValueError(f"Unknown copula type: {self.copula_type}")
